requirements.txt
# .. used in dinogame - enable if playing around with that one...
#matplotlib

#IPython

#jupyter
#notebook
#ipykernel

#pygame

# ^^ used in dinogame - enable if playing around with that one...

numpy
pillow

loguru

dnspython
pyroute2

ollama
google-genai
google-auth
google-auth-oauthlib
google-api-python-client

pydantic
pydantic-settings
pytz
pydantic-extra-types
email-validator

requests

paho-mqtt
schedule

jinja2

PyYAML
ruamel.yaml

cachetools

# piper-tts — installed in /opt/piper-venv via Python 3.12 (piper-phonemize lacks 3.14 wheels)
pathvalidate

mqttstuff==0.0.5
reputils==0.0.16
